refactor(tree): move tree status utils (#17263)
