+++ experimental/dds/tree2/src/feature-libraries/editable-tree/utilities.ts
@@ -6,11 +6,10 @@
 import { assert } from "@fluidframework/core-utils";
 import { isStableId } from "@fluidframework/container-runtime";
 import {
-	DetachedField,
 	FieldKey,
 	TreeStoredSchema,
 	UpPath,
-	keyAsDetachedField,
+	getDetachedFieldContainingPath,
 	rootField,
 } from "../../core";
 import { brand } from "../../util";
@@ -105,23 +104,6 @@ export function getStableNodeKey(
 		);
 		return brand(id);
 	}
-}
-
-/**
- * Checks whether or not a given path is parented under the root field.
- * @param path - the path you want to check.
- * @returns the {@link DetachedField} which contains the path.
- */
-export function getDetachedFieldContainingPath(path: UpPath): DetachedField {
-	let currentPath = path;
-	while (currentPath !== undefined) {
-		if (currentPath.parent === undefined) {
-			return keyAsDetachedField(currentPath.parentField);
-		} else {
-			currentPath = currentPath.parent;
-		}
-	}
-	return keyAsDetachedField(path.parentField);
 }
 
 /**
